CONSOLE-4733: Refactor middleware into dedicated package
Extract middleware from pkg/server/middleware.go into pkg/middleware/
package to improve separation of concerns and align with standard Go
project organization. This refactoring breaks down the monolithic server
package by moving HTTP middleware into a focused, reusable library.

Changes:
- Move middleware from pkg/server to pkg/middleware package
- Update all middleware references in server.go and handler.go
- Export middleware functions (AuthMiddleware, WithBearerTokenReview, etc.)
- Add HTTP method restriction to catalogd metas endpoint
- Fix typo: getCatlogLastModifiedKey → getCatalogLastModifiedKey
- Simplify query parameter encoding in catalog client
